Updated README.md with server startup instructions
In this commit, the README.md file was updated to include instructions on how to start the Django server for the project. This includes the command to run the server and the URL to access the application in a web browser.

This update will make it easier for other developers to get the application running on their local machines.